# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: Restaurant QR Ordering API
  version: 0.1.0
  description: >
    HTTP API for QR-based restaurant ordering. Most staff endpoints require a JWT
    bearer token and a server-resolved permission. Guest endpoints under /api/v1/guest require an
    X-Session-Token issued by /api/v1/dining/join-session. Public endpoints require
    neither.
  x-response-envelope:
    description: Every JSON response includes the numeric HTTP status in `code`.
servers:
  - url: /
...
paths:
...
  /api/v1/identity/authenticate:
    post:
      tags: [identity]
      summary: Login, returns JWT
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthenticateRequest'
            example:
              restaurant_code: DEMO
              username: cashier
              password: demo1234
      responses:
        '200':
          description: Authenticated; response includes permission codes for shell rendering
        '401':
          description: Invalid credentials
  /api/v1/identity/me:
    get:
      tags: [identity]
      summary: Return the current staff identity and permissions
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user, role, restaurant, and permission codes
        '401':
          description: Invalid or expired token
  /api/v1/identity/manage-users:
    post:
      tags: [identity]
      summary: Create or update users
      security:
        - bearerAuth: []
      description: 'Permission: identity.manage'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScaffoldRequest'
      responses:
        '200':
          description: User mutation result
        '403':
          description: Forbidden
  /api/v1/catalog/create-menu-item:
    post:
      tags: [catalog]
      summary: Create menu item
      security:
        - bearerAuth: []
      description: 'Permission: catalog.manage. Endpoint scaffold.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScaffoldRequest'
      responses:
        '200':
          description: Created
  /api/v1/catalog/update-menu-item:
    post:
      tags: [catalog]
      summary: Update menu item
      security:
        - bearerAuth: []
      description: 'Permission: catalog.manage. Endpoint scaffold.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScaffoldRequest'
      responses:
        '200':
          description: Updated
  /api/v1/catalog/delete-menu-item:
    post:
      tags: [catalog]
      summary: Delete menu item
      security:
        - bearerAuth: []
      description: 'Permission: catalog.manage. Endpoint scaffold.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScaffoldRequest'
      responses:
        '200':
          description: Deleted
  /api/v1/catalog/toggle-availability:
    post:
      tags: [catalog]
      summary: Set menu item availability
      security:
        - bearerAuth: []
      description: 'Permission: catalog.manage. Endpoint scaffold.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScaffoldRequest'
      responses:
        '200':
          description: Availability set
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string, format: uuid }
                  status: { type: string }
        '403':
          description: Forbidden (not MANAGER)
  /api/v1/dining/join-session:
    post:
      tags: [dining]
      summary: QR guest entry; issues session token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JoinSessionRequest'
      responses:
        '200':
          description: Session joined
  /api/v1/dining/open-session:
    post:
      tags: [dining]
      summary: Open a table session
      security:
        - bearerAuth: []
      description: 'Permission: dining.serve'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpenSessionRequest'
      responses:
        '200':
          description: Session opened
  /api/v1/dining/close-session:
    post:
      tags: [dining]
      summary: Close a table session
      security:
        - bearerAuth: []
      description: 'Permission: dining.cashier'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CloseSessionRequest'
      responses:
        '200':
          description: Session closed
  /api/v1/dining/manage-table-qr:
    post:
      tags: [dining]
      summary: Manage table QR codes
      security:
        - bearerAuth: []
      description: 'Permission: dining.manage'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManageTableQRRequest'
      responses:
        '200':
          description: QR managed
  /api/v1/dining/table-qrs:
    get:
      tags: [dining]
      summary: List restaurant tables and active QR codes
      security:
        - bearerAuth: []
      description: 'Permission: dining.manage'
      responses:
        '200':
          description: Table QR list
...
  /api/v1/billing/build-invoice:
    post:
      tags: [billing]
      summary: Build an invoice
      security:
        - bearerAuth: []
      description: 'Permission: billing.process'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BuildInvoiceRequest'
      responses:
        '200':
          description: Invoice built
  /api/v1/billing/adjust-invoice:
    post:
      tags: [billing]
      summary: Adjust an invoice
      security:
        - bearerAuth: []
      description: 'Permission: billing.process'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdjustInvoiceRequest'
      responses:
        '200':
          description: Invoice adjusted
  /api/v1/billing/process-payment:
    post:
      tags: [billing]
      summary: Process a payment
      security:
        - bearerAuth: []
      description: >
        Permission: billing.process. Cash/card complete synchronously. Configured
        E_WALLET methods create a PROCESSING payment and return pay_url,
        deeplink, and qr_code_url for gateway checkout; the invoice is paid only
        after the provider webhook is reconciled.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProcessPaymentRequest'
      responses:
        '200':
          description: Payment processed or initiated
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    sessionToken:
      type: apiKey
      in: header
      name: X-Session-Token
  schemas:
    ScaffoldRequest:
      type: object
      description: Endpoint scaffold. The backend currently accepts an empty JSON object only.
      additionalProperties: false
      example: {}
...
    AuthenticateRequest:
      type: object
      required: [restaurant_code, username, password]
      properties:
        restaurant_code: { type: string, example: DEMO }
        username: { type: string, example: cashier }
        password: { type: string, format: password, example: demo1234 }
    JoinSessionRequest:
      type: object
      required: [qr_token]
      properties:
        qr_token: { type: string }
    OpenSessionRequest:
      type: object
      required: [table_id]
      properties:
        table_id: { type: string, format: uuid }
    CloseSessionRequest:
      type: object
      required: [session_id]
      properties:
        session_id: { type: string, format: uuid }
    ManageTableQRRequest:
      type: object
      required: [table_id]
      properties:
        table_id: { type: string, format: uuid }
        rotate: { type: boolean, default: false }
...
    BuildInvoiceRequest:
      type: object
      required: [dining_session_id]
      properties:
        dining_session_id: { type: string, format: uuid }
    AdjustInvoiceRequest:
      type: object
      required: [invoice_id, discount_amount_vnd]
      properties:
        invoice_id: { type: string, format: uuid }
        discount_amount_vnd: { type: integer, format: int64, minimum: 0 }
        discount_reason:
          type: string
          enum: [promo, regular, complaint]
    ProcessPaymentRequest:
      type: object
      required: [invoice_id, payment_method_code]
      properties:
        invoice_id: { type: string, format: uuid }
        payment_method_code:
          type: string
          example: cash
        received_amount_vnd:
          type: integer
          format: int64
          example: 500000
        reference_code: { type: string }
...